+++ apps/formio-server/config.js
@@ -45,6 +45,7 @@ config.formio.reservedForms = [
   'report',
   'version',
   'tag',
+  'owner',
   'exists',
   'export',
   'import',
@@ -194,6 +195,9 @@ config.logging = {
   console: process.env.LOGGING_CONSOLE || true,
   formio: process.env.LOGGING_FORMIO || false
 };
+
+// Remote Auth
+config.remoteAuth = (process.env.REMOTE_AUTH === 'false') ? false : (process.env.REMOTE_AUTH || 'https://api.form.io');
 
 // Allow the config to be displayed when debugged.
 var sanitized = _.clone(config, true);