+++ apps/formio-server/serverSetup.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-npm install
-git submodule update --init --recursive
-cd node_modules/formio
-npm install
-cd ../..