+++ apps/formio-server/src/db/updates/index.js
@@ -9,6 +9,6 @@ module.exports = {
   '3.0.6-rc.1': require('./3.0.6-rc.1.js'),
   '3.0.7-rc.1': require('./3.0.7-rc.1.js'),
   '3.0.8-rc.1': require('./3.0.8-rc.1.js'),
-  '3.0.9-rc.1': require('./3.0.9-rc.1.js')
-  //'3.1.0-rc.1': require('./3.1.0-rc.1.js')
+  '3.0.9-rc.1': require('./3.0.9-rc.1.js'),
+  '3.1.0-rc.1': require('./3.1.0-rc.1.js')
 };