+++ apps/formio-server/src/resources/TagResource.js
@@ -33,16 +33,24 @@ module.exports = function(router, formioServer) {
       formio.middleware.tagHandler,
       function(req, res, next) {
         cache.loadCurrentProject(req, (err, project) => {
-          let options = router.formio.formio.hook.alter('exportOptions', {}, req, res);
-          formio.template.export(options, function(err, template) {
-            if (err) {
-              return res.status(400).send(err);
-            }
-            template.tag = req.body.tag;
-            req.body.template = template;
+          // Allow passing template from frontend. This is useful for remote environments.
+          if (!req.body.template) {
+            let options = router.formio.formio.hook.alter('exportOptions', {}, req, res);
+            formio.template.export(options, function(err, template) {
+              if (err) {
+                return res.status(400).send(err);
+              }
+              template.tag = req.body.tag;
+              req.body.template = template;
+              req.body.owner = project.owner.toString();
+              return next();
+            });
+          }
+          else {
+            req.body.template.tag = req.body.tag;
             req.body.owner = project.owner.toString();
             return next();
-          });
+          }
         });
       }
     ],
@@ -156,7 +164,37 @@ module.exports = function(router, formioServer) {
             if (!('environment' in deploy)) {
               return res.status(400).send('Deploy environment command must contain an environment name.');
             }
+            res.status(400).send('Deploy environment not yet supported');
             // TODO: do environment deploy.
+            break;
+          }
+          case 'template': {
+            if (!('template' in deploy) || typeof deploy.template !== 'object') {
+              res.status(400).send('Must send a template with a template deployment.');
+            }
+
+            let template = deploy.template;
+
+            Object.assign(template, _.pick(project, ['name', 'title', 'description', 'machineName']));
+            let alters = hook.alter('templateAlters', {});
+            debug('import template', template);
+
+            formio.template.import.template(template, alters, function(err, template) {
+              if (err) {
+                debug(err);
+                return res.status(400).send('An error occurred with the template import.');
+              }
+
+              project.tag = template.tag;
+              project.markModified('tag');
+              project.save((err) => {
+                if (err) {
+                  return res.status(400).send(err.message || err);
+                }
+                return res.send('Tag Deployed');
+              });
+            });
+
             break;
           }
           default: {